Start a new session for the user with an expired session key (issue #6)
The web module should be simple. Therefore, it was split on a few modules
which take care of one specific thing. Also, the web part working with sessions
don't start a new session. It is a responsibility of the Sirko.Session module.

Session keys assigned to users through cookies get expired in 1 hour. The idea is to
expire session keys in browsers and sessions in the DB at the same time.