Remove stale data (issue #21)
Stale transitions are useless once links between pages get changed. Actually, they are harmful as they lead to wrong predictions.

To avoid such kind of an issue, there are a few important changes:

 - all session which are older than 7 days get removed
 - transitions with 0 count get removed
 - pages which don't have any relations get removed (there is a chance that those pages get removed from a site as well)

To implement the logic described above, the Sirko.Scheduler.Server was renamed to Sirko.Scheduler.Worker.
Now it is decoupled from the code expiring sessions and it can be used for periodical running of any code.