optimize the solution for expiring inactive sessions (issue #24)
Neo4j performs better when it doesn't get bumped with lots of queries at the same
time (N+1 query problem). Therefore, queries related to the solution which expire sessions
and update transitions were altered to work with a list of sessions.
It resulted in having less queries to Neo4j and drastic performance improvement.
Now it takes ~1 sec to expire 1000 sessions with 10 visited pages.
Before this change, it took ~35 secs to do the same job.