track active users who typed the direct url to a page (#27)
When an active user types a url, we don't track anything right away. But, when a user
goes from that page to another one, we must track the transition. Therefore,
even if a reffer is a new to the engine, it gets created in the DB.

Initially, the goal was to get a smooth navigation path. But, users
may type urls or get url hints from the browser or use bookmarks whenever
they want, hence, the smooth transition isn't fully reached.