introduce a confidence threshold setting (#32)
The prerender resource hint increases the load on a backend. Therefore, before adding the prerender hint
we need to be confident enough that the current user will visit the predicted page. In order to leverage that confidence,
a new 'sirko.engine.confidence_threshold' was introduced. Sites which suffer from high load because of the prerender hint
can define higher value to reduce the load. Meantime, static sites might stick to lower value.