no missing transitions anymore (#29) & remove the starting point (#33)
The starting point didn't bring any value to the model. Users can start exploring a site from
different pages. The starting point was useful for keeping a newly generated session key of a new session.
But, another solution was found. Now when a user visits a page, the session key gets generated and added to the DB only
after the first transition. So, we do it when we have the refer.