add an index to the Page node (#36)
Indexes are added on start of the application in a separate process to not effect the launch.

The distillery provides boot hooks which could be used for adding indexes, but it would only work in the packaged app. So, if that hook got broken after upgrading deps/elixir, we would spot it only after packaging. It is better to spot such issues in development.