store and predict assets (#22)
Now the engine accepts a list of assets of a page. That list gets
stored along with the path of the page. Once the prediction is made,
the path and assets of a predicted page gets returned in the response.
So, the client can instruct browsers to prefetch the predicted page and
assets of it.

Since the engine needs to work with the more complex structure of data,
it got altered to accept requests in the json format and respond
in this format too. In particular, we could use a pure POST request,
but the decision was made to use the same format for requests
and responses in order to be consistent.