predict several pages (sirko-io/client#29)
Some sites might be ok to prefetch a few pages which might
be visited by the current user. Thus, a chance that the next
page will be served from the cache gets higher.

Now, the engine might predict a few pages. A max number of
pages is controlled via a `max_pages_in_prediction` config value.
The confidence threshold is still considered, it means predicted
pages must pass it to be included into the response.